Squash all NPM into one service
