+++ Jvw.DevToys.SemverCalculator/Detectors/SemVersionRangeDataTypeDetector.cs
@@ -7,11 +7,11 @@ namespace Jvw.DevToys.SemverCalculator.Detectors;
 /// <summary>
 /// Data-type detector for DevToys that can detect SemVer values and ranges.
 /// </summary>
-/// <param name="packageVersionService">Package version service.</param>
+/// <param name="packageManagerService">Package manager service.</param>
 [Export(typeof(IDataTypeDetector))]
 [DataTypeName(Name, baseName: PredefinedCommonDataTypeNames.Text)]
 [method: ImportingConstructor]
-internal sealed class SemVersionRangeDataTypeDetector(IPackageVersionService packageVersionService)
+internal sealed class SemVersionRangeDataTypeDetector(IPackageManagerService packageManagerService)
     : IDataTypeDetector
 {
     internal const string Name = "SemVersionRange";
@@ -26,7 +26,7 @@ internal sealed class SemVersionRangeDataTypeDetector(IPackageVersionService pac
         if (
             rawData is string dataString
             && !string.IsNullOrEmpty(dataString)
-            && packageVersionService.IsValidRange(dataString)
+            && packageManagerService.IsValidRange(dataString)
         )
         {
             return ValueTask.FromResult(new DataDetectionResult(Success: true, dataString));
