+++ Jvw.DevToys.SemverCalculator/Services/NpmService.cs
@@ -4,25 +4,29 @@ using System.Diagnostics.CodeAnalysis;
 using System.Net;
 using System.Text.Json;
 using DevToys.Api;
+using Jvw.DevToys.SemverCalculator.Enums;
 using Jvw.DevToys.SemverCalculator.Models;
 using Microsoft.Extensions.Logging;
+using Semver;
 
 namespace Jvw.DevToys.SemverCalculator.Services;
 
 /// <summary>
 /// Fetch package versions from the NPM registry.
 /// </summary>
-[Export(typeof(INpmService))]
-internal class NpmService : INpmService
+[Export(typeof(IPackageManagerService))]
+internal class NpmService : IPackageManagerService
 {
-    /// <summary>
-    /// JSON serializer options.
-    /// </summary>
     private readonly JsonSerializerOptions _jsonSerializerOptions =
         new() { PropertyNameCaseInsensitive = true };
 
     private readonly HttpClient _httpClient;
     private readonly ILogger _logger;
+    private List<SemVersion> _versions = [];
+    private SemVersionRange? _range;
+
+    /// <inheritdoc cref="IPackageManagerService.PackageManager" />
+    public PackageManager PackageManager => PackageManager.Npm;
 
     /// <summary>
     /// Fetch package versions from the NPM registry.
@@ -45,7 +49,54 @@ internal class NpmService : INpmService
         _logger = logger ?? this.Log();
     }
 
-    /// <inheritdoc cref="INpmService.FetchPackage" />
+    /// <inheritdoc cref="IPackageManagerService.SetVersions" />
+    public void SetVersions(List<string> versions)
+    {
+        _versions = versions.Select(v => SemVersion.Parse(v, SemVersionStyles.Strict)).ToList();
+        _versions.Sort(SemVersion.SortOrderComparer);
+    }
+
+    /// <inheritdoc cref="IPackageManagerService.GetVersions" />
+    public IEnumerable<(string version, bool match)> GetVersions(bool includePreReleases)
+    {
+        foreach (var version in _versions)
+        {
+            if (!includePreReleases && version.IsPrerelease)
+            {
+                continue;
+            }
+
+            yield return (version.ToString(), _range != null && _range.Contains(version));
+        }
+    }
+
+    /// <inheritdoc cref="IPackageManagerService.TryParseRange" />
+    public bool TryParseRange(string value)
+    {
+        var valid = TryParseRange(value, out _range);
+        if (!valid)
+            _range = null;
+        return valid;
+    }
+
+    /// <inheritdoc cref="IPackageManagerService.IsValidRange" />
+    public bool IsValidRange(string value)
+    {
+        return TryParseRange(value, out _);
+    }
+
+    /// <summary>
+    /// Try to parse range.
+    /// </summary>
+    /// <param name="value">Range value.</param>
+    /// <param name="versionRange">Parsed version range.</param>
+    /// <returns>Whether value is valid range.</returns>
+    private static bool TryParseRange(string value, out SemVersionRange versionRange)
+    {
+        return SemVersionRange.TryParseNpm(value, true, out versionRange);
+    }
+
+    /// <inheritdoc cref="IPackageManagerService.FetchPackage" />
     public async Task<PackageJson?> FetchPackage(string packageName)
     {
         _logger.LogInformation("Fetching package \"{PackageName}\"...", packageName);
